requirements.txt
# List of used packages
control==0.9.4
cvxpy==1.2.0
imageio==2.9.0
matplotlib==3.3.4
numpy==1.20.1
imageio-ffmpeg==0.4.8
pandas==1.3.2
progressbar2==3.55.0
pycddlib==2.1.6
pyopengl==3.1.5
pyqt5==5.15.2
scipy==1.6.2
seaborn==0.11.2
visvis==1.12.4
xlrd==2.0.1
xlsxwriter==3.0.1
